refactor: extract plan hooks
